% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields}
\alias{fields}
\title{Possible Fields to Return from Place Details}
\format{
An object of class \code{character} of length 40.
}
\source{
\url{https://developers.arcgis.com/rest/places/place-id-get/#requestedfields}
}
\usage{
fields
}
\description{
A character vector containing the possible return fields that define the attributes to return. Use in \code{\link[=place_details]{place_details()}}'s \code{requested_fields} argument.
}
\keyword{datasets}
