% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-near-point.R
\name{near_point_ptype}
\alias{near_point_ptype}
\title{Create a prototype data.frame for the near-point query}
\usage{
near_point_ptype()
}
\description{
Create a prototype data.frame for the near-point query
}
\keyword{internal}
