% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_category_details}
\alias{parse_category_details}
\title{Parse \verb{/categories/\{categoryId\}} results vectorized}
\usage{
parse_category_details(x)
}
\description{
Parse \verb{/categories/\{categoryId\}} results vectorized
}
\keyword{internal}
