% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisplaces-package.R
\docType{package}
\name{arcgisplaces-package}
\alias{arcgisplaces}
\alias{arcgisplaces-package}
\title{arcgisplaces: Search for POIs using ArcGIS 'Places Service'}
\description{
The ArcGIS 'Places service' is a ready-to-use location service that can search for businesses and geographic locations around the world. It allows you to find, locate, and discover detailed information about each place. Query for places near a point, within a bounding box, filter based on categories, or provide search text. 'arcgisplaces' integrates with 'sf' for out of the box compatibility with other spatial libraries.
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
